Modify "ingredients" heading and add recipe steps
The "ingredients" heading did not specify that the ingredients list was for for Tex-Mex Burgers. This commit adds this specification to the heading so that users can measure ingredient quantities accurately when deciding the number of burgers to make. The webpage did not have a step by step recipe guide. This commit adds this feature via a "Steps" heading and list. Users should now be able have a recipe guide for using the ingredients. Paragraph tags were placed between each "Steps" list tag to create a space between the lists on the webpage.