Add pizza.html and a link in index.html
A recipe page for mexican pizza was needed. A link to the page from homepage was alo needed. This commit adds this feature to the website. This was done by creating an HTML boilerplate called pizza.html and adding its relative link to the index.html file. It was done this way because it is most convenient.